pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "keel-cli"
version = "0.1.0"
description = "KEEL is a local-first anti-drift CLI for understanding and guiding work in messy codebases."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "typer>=0.12",
  "pydantic>=2.8",
  "rich>=13.7",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
keel = "keel.cli.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
